broken(keymgmt): add encapsulate_ex method to KeyPair
Added a new method `encapsulate_ex` to the `KeyPair` implementation. This
method is intended to encapsulate a key and return an `EncapsulatedKey`
and `SharedSecret`.
The goal is for this convenience method to use the `rng` from the
`provctx`.

Currently, the method is marked with `todo!()` and
includes a trace log statement.