test(adapters): disable libcrux adapter initialization
Commented out the initialization of the `libcrux` adapter and marked
its module with `#[expect(dead_code)]` to suppress warnings.

This change temporarily disables `libcrux` while keeping its code
intact.

- Added `#[expect(dead_code)]` to `libcrux` to suppress warnings.
- Commented out `libcrux::init()` to prevent its initialization.

This change is to test the ability to easily disable a specific adapter
by just removing the call to its `init` function.

As expected, applying this change we are unable to connect to Cloudflare
as the `libcrux_draft` adapter does not register algs/caps compatible
with their server!

Local loopback connection with the remaining supported group works.