chore(dependencies): cargo update
There are issues with the versions of rand and rand_core that get pulled
in by the 0.0.2 versions of the various libcrux crates (of which there
are many, most of which we don't explicitly depend on), so for now I
went through and did a `cargo update --recursive <package>` for each of
the *other* packages listed as explicit dependencies in Cargo.toml
(function_name, libc, log, etc. -- all of our explicit dependencies
except openssl_provider_forge and libcrux-kem).