refactor(encoder): use the transcoders module from forge with dedicated Encoder trait
Requires that the forge crate be at v0.8.4-dev @ 47c204fe or newer.

Tests are going to fail until this crate is updated to depend on
a sufficient version of the forge crate, but those commits aren't on
the official GitHub repository yet.

Switched openssl_provider_forge dependency to use the pre-release
0.8.4-dev version from romen's repository at `47c204fe`.

The previous dependency from the upstream qubip repository (tag v0.8.3)
is now commented out.