refactor(encoders): use a macro to generate plain text encoders for public keys
This is also a minor bugfix: the encoder now includes a terminating null
byte when writing out the string. The lack of one wasn't causing any
visible problems before, but since we're writing a string to memory for
a C application to use, I'm fairly certain it should be there.