fix(tests): check test flags before describing key decoding error as "expected"
Previously, if decoding of the pubkey or privkey failed, the ML-DSA
Wycheproof tests would report that the attempt to decode the key "failed
as expected" (and mark the test as passed) for every test case that was
meant to fail, including test cases that were meant to successfully
decode the key but then fail for some other reason (e.g. the "context
too long" error expected in test case 5 of the signing tests).

(In practice, this only showed up for the privkey in the signing tests,
since tc5 was unexpectedly passing when I started running the seed-only
signing tests which aurora shouldn't yet be able to pass any of, but
I've preemptively fixed the same type of problem for the pubkey in the
verification tests as well.)